Eros Q Native port?
https://thesciencedog.files.wordpress.com/2013/09/golden-retriever-and-science1.jpg

Input/Changes very much welcome

We can get to rockbox.bin via usb boot, but the LCD is screwed up somehow.
It passes initialization (almost guaranteed I've got bit depth and format wrong tho),
but we get stuck waiting for end-of-frame via LCD_STATE register for some reason.

This is (I think) set up for PIO mode for the LCD. However, all attempts
to get the LCD to respond have not been fruitful.

This is messy.

Change-Id: I70dda8fc092c6e3f4352f2245e4164193f803c33