Update patch set 8
Patch Set 8:

Can you describe how/when lcd_dma_wr_cmd[] is used? Is it supposed to contain a command to tell x1000 to do dma, or is it supposed to contain my command to the LCD to start the frame data? I'm having a hard time following where/how it's used. 

My suspicion is that it's the LCD_MDATA 0x2c (w/ ptr set to 0x00) that was the magic in PIO mode, but... in that case shouldn't it need to be { 0x00, 0x00, 0x00, 0x2c } like on the m3k? or is each entry sent on its own? (why the length of 4?)

Patch-set: 8