Update patch set 10
Patch Set 10:

> I see... The only thing I don't understand is, isn't lcd_update_rect() a partial redraw? Or is that function not really used? Or... does it put the data in the framebuffer and then we wait until a full frame is ready before sending it?

The Rockbox API forces us to assume *every* update call might stand alone and the result needs displaying to the user. lcd_update_rect() is a partial redraw, but because we only transfer full frames, we can only optimize the copy from the main framebuffer to our local "shadow framebuffer." Copying from the main FB to shadow FB is accelerated by DMA.

The full shadow FB is always transferred to the panel controller using DMA in the background. The transfer is initiated by any lcd_update*() call (partial or not) and runs asynchronously so the call can return quickly.

If updates occur too fast, we need to wait for the previous frame transfer to finish before we can copy the next update into the shadow FB (to avoid tearing). That is why many small partial updates are a bad case -- most of the time we get stuck waiting for the previous small update's frame to finish.

Patch-set: 10