Update patch set 15
Patch Set 15:

All you need to do for audio is set up the clocks and power on the codec. It doesn't have any fancy settings so it should just be a matter of setting the GPIOs to the right level.

The X1000's audio clocks are generated like this:

    (PLL source) --> system clock --> bit clock --> frame clock

The frame clock equals the sample rate, bit clock is always 64fs, and the system clock is generated from a PLL using CPM_I2SCDR. I2SCDR is a fractional divider but the bit clock is generated with a simple 1/N divider.

Due to the fixed 64fs bit clock, the X1000 always outputs 32-bit stereo samples. The default format is I2S but MSB-justified is supported too.

DMA transfer of audio data to the FIFO is handled by pcm-x1000.c and shouldn't need to be changed.

Patch-set: 15