Time and date format unification (Igor Poretsky)
More general pronunciation style for numeric values, time and date to
make it better fit for various languages.

Time and date related strings for display and voice and time format
option have been made available on non-RTC targets as well, since now
they are used not only for current date and time. Date is now spoken
with the weekday if available.

Change-Id: I667350e85015a597da572b93487c683c84922a85