Update patch set 21
Patch Set 21:

> Patch Set 21:
> I'm fairly sure the column address width was the only problem. I doubt the remapping had anything to do with it. The x-loader code (drivers/lpddr.c) suggests the remapping helps spread memory access across multiple banks in DRAM. Probably there is a performance benefit to this, similar to how having multiple DIMMs improves performance; I don't know enough about DRAM or memory controllers to say for sure.
Sounds good, I'll go for the original values and see what happens.
> Another thing -- the clock setup has been moved to clk-x1000.c, so the 2nd PLL (MPLL) will be left enabled at 600 MHz unless you add the ErosQ to the M3K's ifdef path. The PLL may draw an extra 0.5 mA or so, but the exact increase was too small to measure when I did some limited testing on the Q1. (I'm thinking of using the 2nd PLL on the Q1 because its LCD is finicky about clock rates.)
> 
> On the M3K there's no problems and no reason to switch configs so I kept the APLL-only setup. That is the one you've been developing with, so if you haven't found any weird issues on the ErosQ you may want to save the tiny bit of power by following the M3K's setup.
Ok, I'll give it a try today I think, hopefully I can get it rebased and working without much trouble. On the hosted version, the LCD does sometimes flicker when you wake it up after it's been off, but I don't think I've seen that on this version. Then again, if it does show up later that's always an option to try to get rid of that bug.

> On another note, how is this coming feature-wise? Anything major missing, instabilities, etc? Ready for review or does it still need work? (I know the bootloader is missing but I'm refactoring the M3K stuff into a generic X1000 bootloader/installer and it should work here too.)

I think it's getting close. Honestly, I don't think I've really had /any/ outright crashes (especially recently), so I guess it seems pretty stable. We're missing a bootloader obvs, but beyond that I would like to try to brute-force find the missing DAC control lines if they're hooked up, since those pins are big enough I can get my multimeter on them. I also need to re-enable the plugins and see what that situation looks like. 

I don't have any information about the bluetooth module, but I don't think that's very important since we don't have support for BT anyway so I won't mess with that. If it shows to draw a crazy amount of power without intervention that could be fixed later.

And, of course the volume scaling issue, but that can wait and be its own thing.

Someday I would like to try getting the LCD working with a higher bit depth, but it works well as it is so I'll leave it be for now.

I'd be lying if I said I wasn't quite excited about getting this "done". I'm sure it will need a lot of cleaning up in review to get it to merge-able quality code though 😄

Patch-set: 21