Update patch set 1
Patch Set 1:

Re. the proliferation of options & defines: we could define settings externally to the settings code, in a target-specific file. Such as a static table which the settings code would point to and use, so the contents of the table don't have to physically go into settings_list.c.

Codecs and maybe a handful of other very target-specific settings could benefit from this.

I don't know what the settings code implementation is like, so I have no idea if this is feasible. Or maybe there's some problem I'm not realizing here. Any thoughts?

Patch-set: 1
CC: Gerrit User 1000718 <1000718@583131e1-1a33-4f2b-b8ec-82434880f207>