Update patch set 24
Patch Set 24:

Here's another question - I'm reading through the X1000 programming manual, do you think the internal codec has any value to us? Or is it even possible to use it as we're set up now? It appears to have quite a few gain/mixer controls, and if I am reading it right, it can interface with AIC to run an I2S DAC.

Patch-set: 24