Update patch set 24
Patch Set 24:

(5 comments)

> Patch Set 24:
> 
> (7 comments)
> 
> Some initial review. It's basically OK. There's still a handful of spurious changes (like files with blank lines being added) which should be fixed. I'll give it a 2nd more careful pass once you have the bootloader working. Let me know if there is anything else you want to get done before merging.

Thanks for the review - I'll get the stuff you mentioned sorted out.

One thing I'm not sure about is the naming of files - I seem to recall when you did the m3k you renamed everything m3k (linux) to something like fiiom3klinux. Does that need to be done? I want to follow convention, but I'm also hesitant to rename a bunch of stuff that's working just fine.

I'm also unsure about the naming of the pcm5102a.c file - do you think it should be something like "erosqnative_codec.c" or something like that? The only weird thing is the existing *_codec.c files seem to be limited to the hosted stuff. There's really not much to it - in fact, the stuff that's there is only there because I needed access to the settings, and for some reason couldn't get that in audiohw_erosqnative.c.

Patch-set: 24