Update patch set 26
Patch Set 26:

So I have the bootloader working with usbboot - I've tested all the functions that don't write to nand:

- Start Rockbox
- USB mode (btw, for some reason this is super cool to me!)
- Shutdown
- Reboot
- Bootloader Backup

The bootloader backup gave me some trouble until I remembered you said I needed to add the erosq to the #ifdef that defines the nand chip, then it let me dump the contents.

I would like to try writing this bootloader to nand (nervous...), is there anything I need to know or do before I just go for it? Do I need to examine the nand dumps I did earlier to find any information or anything? Need to define an address or something?

Presumably, since I already took a backup of the full 128M of the nand, it's possible to write that backup back to it should something go wrong, correct?

Patch-set: 26