Update patch set 26
Patch Set 26:

Aidan, thanks for your hard work - it looks like it works great! I did discover that the Vol+ button is unavailable for the recovery mode button - it appears to have some hardware function upon boot, as it just turns on to a black screen. So I went with Vol-, and all functions (minus dual boot, obvs) seem to work without issue. The scrollwheel even works in the recovery menu - neat!

> Patch Set 26:
> I can explain a lot more about the boot process and what you need for dual boot but I think I should probably add that to the wiki in some general X1000 port information (it would be a _lot_ to put here).
That would probably be a great resource if you wanted to write that up 👍

> And IMHO it's fine to implement dual boot later, in another patch.
I agree - I personally have no intentions of using the OF on this device (it's.... really bad.), but I understand people like to use the USB DAC function, so probably should implement it sooner or later.

Patch-set: 26