Fiio M3K: Fix Yes/No Keymap
The yes/no keymap does not fall back to the Standard
keymapping, so keys other than Back and Power do not
cancel, though the prompt says "any other key".

Adding fallback to standard context as well as explicit
Left, Right, Vol+ and Vol- as they're not present in
the standard context.

Tested on physical m3k

Change-Id: I06409bad8dcde6586e1bf773309ef0ce464cc060