Revert "AXP PMU rewrite (again)"
This caused LCD problems on the ErosQ, where the screen would
go white until being put through a sleep/wake cycle. The exact
reason for this isn't obvious, but the problem didn't exist
prior to the AXP driver rewrite.

The two dependent changes,

42999913ba - x1000: Increase USB current limit to 500 mA at all times
90dd2f84a9 - x1000: Correctly limit USB charging current

ended up bringing the USB charging situation back to where it
was prior to the rewrite, so the cleanest option is to revert
the whole lot.

This reverts commit 42999913ba3a76221fceb04b1f935ed4e0e71476.
This reverts commit 90dd2f84a9174c38dbfb07d582ec6ee7697b1939.
This reverts commit 2d891439623bb76d38b98202ca5f3eea3c01c5f0.

Change-Id: I1cff2bfdd1b189df14bcf8cce42db725caa470d7